# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Credit History Analyzer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResponse {
  sessionId: string;
  filesCount: number;
  message: string;
}

export type AnalysisResultRating =
  (typeof AnalysisResultRating)[keyof typeof AnalysisResultRating];

export const AnalysisResultRating = {
  excellent: "excellent",
  good: "good",
  average: "average",
  poor: "poor",
  bad: "bad",
} as const;

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Recommendation {
  title: string;
  description: string;
  priority: RecommendationPriority;
}

export type RiskSeverity = (typeof RiskSeverity)[keyof typeof RiskSeverity];

export const RiskSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Risk {
  title: string;
  description: string;
  severity: RiskSeverity;
}

export interface AnalysisResult {
  sessionId: string;
  rating: AnalysisResultRating;
  ratingScore: number;
  ratingLabel: string;
  summary: string;
  totalDebts: number;
  activeLoans: number;
  closedLoans: number;
  overdueLoans: number;
  debtBurdenRatio: number;
  inquiriesLastMonth: number;
  recommendations: Recommendation[];
  risks: Risk[];
  reportReady: boolean;
}

export type SessionStatusStatus =
  (typeof SessionStatusStatus)[keyof typeof SessionStatusStatus];

export const SessionStatusStatus = {
  uploading: "uploading",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface SessionStatus {
  sessionId: string;
  status: SessionStatusStatus;
  filesCount: number;
  createdAt: string;
  analysisResult?: AnalysisResult;
  errorMessage?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UploadCreditFilesBody = {
  files: Blob[];
};