# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Credit History Analyzer
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analysis
    description: Credit history analysis operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analysis/upload:
    post:
      operationId: uploadCreditFiles
      tags: [analysis]
      summary: Upload PDF credit history files
      description: Upload one or more PDF files from different credit bureaus for analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - files
      responses:
        "200":
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Invalid files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analysis/sessions/{sessionId}/analyze:
    post:
      operationId: analyzeSession
      tags: [analysis]
      summary: Run AI analysis on uploaded files
      description: Extracts data from PDFs, merges them, runs AI analysis and returns result
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analysis/sessions/{sessionId}/report:
    get:
      operationId: downloadReport
      tags: [analysis]
      summary: Download PDF report
      description: Download the generated PDF report for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analysis/sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [analysis]
      summary: Get session status
      description: Get the current status and details of an analysis session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UploadResponse:
      type: object
      properties:
        sessionId:
          type: string
        filesCount:
          type: integer
        message:
          type: string
      required:
        - sessionId
        - filesCount
        - message

    AnalysisResult:
      type: object
      properties:
        sessionId:
          type: string
        rating:
          type: string
          enum: [excellent, good, average, poor, bad]
        ratingScore:
          type: integer
        ratingLabel:
          type: string
        summary:
          type: string
        totalDebts:
          type: number
        activeLoans:
          type: integer
        closedLoans:
          type: integer
        overdueLoans:
          type: integer
        debtBurdenRatio:
          type: number
        inquiriesLastMonth:
          type: integer
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
        reportReady:
          type: boolean
      required:
        - sessionId
        - rating
        - ratingScore
        - ratingLabel
        - summary
        - totalDebts
        - activeLoans
        - closedLoans
        - overdueLoans
        - debtBurdenRatio
        - inquiriesLastMonth
        - recommendations
        - risks
        - reportReady

    Recommendation:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
      required:
        - title
        - description
        - priority

    Risk:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [high, medium, low]
      required:
        - title
        - description
        - severity

    SessionStatus:
      type: object
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [uploading, processing, completed, failed]
        filesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        analysisResult:
          $ref: "#/components/schemas/AnalysisResult"
        errorMessage:
          type: string
      required:
        - sessionId
        - status
        - filesCount
        - createdAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message