# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Credit History Analyzer
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultRating } from "./analysisResultRating";
import type { Recommendation } from "./recommendation";
import type { Risk } from "./risk";

export interface AnalysisResult {
  sessionId: string;
  rating: AnalysisResultRating;
  ratingScore: number;
  ratingLabel: string;
  summary: string;
  totalDebts: number;
  activeLoans: number;
  closedLoans: number;
  overdueLoans: number;
  debtBurdenRatio: number;
  inquiriesLastMonth: number;
  recommendations: Recommendation[];
  risks: Risk[];
  reportReady: boolean;
}