# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Credit History Analyzer
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from "./recommendationPriority";

export interface Recommendation {
  title: string;
  description: string;
  priority: RecommendationPriority;
}