# lib/api-zod/src/generated/types/risk.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Credit History Analyzer
 * OpenAPI spec version: 0.1.0
 */
import type { RiskSeverity } from "./riskSeverity";

export interface Risk {
  title: string;
  description: string;
  severity: RiskSeverity;
}