# lib/api-zod/src/generated/types/sessionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Credit History Analyzer
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";
import type { SessionStatusStatus } from "./sessionStatusStatus";

export interface SessionStatus {
  sessionId: string;
  status: SessionStatusStatus;
  filesCount: number;
  createdAt: Date;
  analysisResult?: AnalysisResult;
  errorMessage?: string;
}